operation strategies for add/sub/divide/multiply

+++ TeamProject/src/application/OperationStrategy.java
@@ -2,6 +2,6 @@ package application;
 
 public interface OperationStrategy {
 	
-	public long doOperation(long leftOperand, long rightOperand);
+	public double doOperation(double leftOperand, double rightOperand);
 	
 }
